+++ src/resources/views/review.blade.php
@@ -9,13 +9,7 @@
 @endsection
 
 @section('content')
-<?php
-    // $reviews = App\Models\Review::where('restaurant_id', $id)->get();
-    // $score_avg = $reviews->avg('score');
-    // $count = $reviews->count();
-    // $restaurant_name = $reviews->first()->restaurant->name;
-    // $today = Carbon\Carbon::now()->format('Y-m-d');
-?>
+
 <article class="review-main" id="review-main">
     <section class="review-title">
         <dl class="review-title__restaurant">
@@ -30,9 +24,12 @@
     <section class="review-form-input">
         <dl class="review-form-input__form">
         <form action="/user/review/reviewAdd" method="post">
+            @csrf
+            <input type="hidden" name="restaurant_id" value="{{$restaurant_id}}">
+            <input type="hidden" name="user_id" value="{{Auth::id()}}">
             <dt class="review-form-input__form-visited-date">あなたの訪問日：</dt>
             <dd id="review-form-input__form-visited-date">
-                <input type="date" name="visited-date" max="{{$today}}">
+                <input type="date" name="visited_date" max="{{$today}}">
             </dd>
             <dt class="review-form-input__form-score">あなたの評価点：</dt>
             <dd id="review-form-input__form-score">
@@ -46,34 +43,43 @@
             <dd id="review-form-input__form-comment">
                 <textarea name="comment" id="review-form-input__form-comment" cols="30" rows="10"></textarea>
             </dd>
+            <div class="review-form-input__submit">
+                <input type="submit" value="投稿する">
+            </div>
         </form>
         </dl>
     </section>
     <section class="review-reviews">
+        @if($reviews->count() > 0)
         @foreach($reviews as $review)
-        <dl id="review-reviews-body{{$loop->iteration}}">
-            <dt class="review-reviews-body__review-score">点数：</dt>
-            <dd id="review-reviews-body__review-score">
-                {{$review->score}}
-            </dd>
-            <dt class="review-reviews-body__review-user">投稿者：</dt>
-            <dd id="review-reviews-body__review-user">
-                {{$review->user->name}}
-            </dd>
-            <dt class="review-reviews-body__review-created-at">投稿日：</dt>
-            <dd id="review-reviews-body__review-created-at">
-                {{$review->created_at}}
-            </dd>
-            <dt class="review-reviews-body__review-visited-date">訪問日：</dt>
-            <dd id="review-reviews-body__review-visited-date">
-                {{$review->visited_date}}
-            </dd>
-            <dt class="review-reviews-body__review-comment">コメント：</dt>
-            <dd id="review-reviews-body__review-comment">
-                {{$review->comment}}
-            </dd>
-        </dl>
-        @endforeach
+            <dl id="review-reviews-body{{$loop->iteration}}">
+                <dt class="review-reviews-body__review-score">点数：</dt>
+                <dd id="review-reviews-body__review-score">
+                    {{$review->score}}
+                </dd>
+                <dt class="review-reviews-body__review-user">投稿者：</dt>
+                <dd id="review-reviews-body__review-user">
+                    {{$review->user->name}}
+                </dd>
+                <dt class="review-reviews-body__review-created-at">投稿日：</dt>
+                <dd id="review-reviews-body__review-created-at">
+                    {{$review->created_at}}
+                </dd>
+                <dt class="review-reviews-body__review-visited-date">訪問日：</dt>
+                <dd id="review-reviews-body__review-visited-date">
+                    {{$review->visited_date}}
+                </dd>
+                <dt class="review-reviews-body__review-comment">コメント：</dt>
+                <dd id="review-reviews-body__review-comment">
+                    {{$review->comment}}
+                </dd>
+            </dl>
+            @endforeach
+        @else
+            <h2 class="review-reviews__nothing">レビューがまだありません</h2>
+        @endif            
+        
+        
     </section>
 </article>
 @endsection
